refactor: un-Class-ify the `Room` model & separate scraper's logic out!
